services/itinerary-service-json/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-jose
